Move image save functions to gaphor.diagram.export
They're imported in the plugin so we do not break backwards compatibiity.
