--- gaphor/plugins/diagramexport/tests/conftest.py
@@ -1,6 +1,0 @@
-from gaphor.diagram.tests.fixtures import (
-    diagram,
-    element_factory,
-    event_manager,
-    modeling_language,
-)
